Move pkg to internal, change render test to appropriate file path, delete auto-deploy-app
