--- auto-deploy-app/.gitlab-ci.yml
@@ -1,19 +1,0 @@
-image: registry.gitlab.com/charts/alpine-helm
-
-stages:
-  - test
-  - release
-
-lint:
-  stage: test
-  script:
-    - helm lint .
-  except:
-    - master
-
-release-chart:
-  stage: release
-  script:
-    - curl --fail --request POST --form "token=$CI_JOB_TOKEN" --form ref=master https://gitlab.com/api/v4/projects/2860651/trigger/pipeline
-  only:
-    - master
