--- auto-deploy-app/Chart.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-description: GitLab's Auto-deploy Helm Chart
-icon: https://gitlab.com/gitlab-com/gitlab-artwork/raw/master/logo/logo-square.png
-name: auto-deploy-app
-version: 0.2.9
