--- auto-deploy-app/charts/postgresql/Chart.yaml
@@ -1,18 +1,0 @@
-description: Object-relational database management system (ORDBMS) with an emphasis
-  on extensibility and on standards-compliance.
-engine: gotpl
-home: https://www.postgresql.org/
-icon: https://www.postgresql.org/media/img/about/press/elephant.png
-keywords:
-- postgresql
-- postgres
-- database
-- sql
-maintainers:
-- name: swordbeta
-- name: databus23
-name: postgresql
-sources:
-- https://github.com/kubernetes/charts
-- https://github.com/docker-library/postgres
-version: 0.7.1
