--- auto-deploy-app/charts/postgresql/templates/secrets.yaml
@@ -1,19 +1,0 @@
-apiVersion: v1
-kind: Secret
-metadata:
-  name: {{ template "fullname" . }}
-  labels:
-    app: {{ template "fullname" . }}
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
-    release: "{{ .Release.Name }}"
-    heritage: "{{ .Release.Service }}"
-type: Opaque
-data:
-  {{ if .Values.postgresPassword }}
-  postgres-password:  {{ .Values.postgresPassword | b64enc | quote }}
-  {{ else }}
-  postgres-password: {{ randAlphaNum 10 | b64enc | quote }}
-  {{ end }}
-  {{- if .Values.metrics.customMetrics }}
-  custom-metrics.yaml: {{ toYaml .Values.metrics.customMetrics | b64enc | quote }}
-  {{- end }}
