--- auto-deploy-app/charts/postgresql/templates/svc.yaml
@@ -1,26 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ template "fullname" . }}
-  labels:
-    app: {{ template "fullname" . }}
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
-    release: "{{ .Release.Name }}"
-    heritage: "{{ .Release.Service }}"
-{{- if .Values.metrics.enabled }}
-  annotations:
-    prometheus.io/scrape: "true"
-    prometheus.io/port: "9187"
-{{- end }}
-spec:
-  type: {{ .Values.service.type }}
-  ports:
-  - name: postgresql
-    port: {{ .Values.service.port }}
-    targetPort: postgresql
-{{- if .Values.service.externalIPs }}
-  externalIPs:
-{{ toYaml .Values.service.externalIPs | indent 4 }}
-{{- end }}
-  selector:
-    app: {{ template "fullname" . }}
