--- auto-deploy-app/requirements.yaml
@@ -1,5 +1,0 @@
-dependencies:
-  - name: postgresql
-    version: "0.7.1"
-    repository: "https://kubernetes-charts.storage.googleapis.com/"
-    condition: postgresql.enabled
