--- auto-deploy-app/templates/NOTES.txt
@@ -1,5 +1,0 @@
-{{- if .Values.service.enabled -}}
-Application should be accessible at: {{ .Values.service.url }}
-{{- else -}}
-Application will be accessible at: {{ .Values.service.url }} when you deploy stable track.
-{{- end -}}
