--- auto-deploy-app/templates/db-initialize-job.yaml
@@ -1,39 +1,0 @@
-{{- if .Values.application.initializeCommand -}}
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: {{ template "trackableappname" . }}-db-initialize
-  labels:
-    app: {{ template "appname" . }}
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version| replace "+" "_" }}"
-    release: {{ .Release.Name }}
-    heritage: {{ .Release.Service }}
-  annotations:
-    "helm.sh/hook": post-install
-    "helm.sh/hook-delete-policy": before-hook-creation
-    "helm.sh/hook-weight": "0"
-spec:
-  template:
-    metadata:
-      labels:
-        app: {{ template "appname" . }}
-        release: {{ .Release.Name }}
-    spec:
-      restartPolicy: Never
-      imagePullSecrets:
-{{ toYaml .Values.image.secrets | indent 10 }}
-      containers:
-      - name: {{ .Chart.Name }}
-        image: "{{ .Values.image.repository }}:{{ .Values.image.tag }}"
-        command: ["/bin/sh"]
-        args: ["-c", "{{ .Values.application.initializeCommand }}"]
-        imagePullPolicy: {{ .Values.image.pullPolicy }}
-        {{- if .Values.application.secretName }}
-        envFrom:
-        - secretRef:
-            name: {{ .Values.application.secretName }}
-        {{- end }}
-        env:
-        - name: DATABASE_URL
-          value: {{ .Values.application.database_url | quote }}
-{{- end -}}
