--- auto-deploy-app/templates/deployment.yaml
@@ -1,66 +1,0 @@
-{{- if not .Values.application.initializeCommand -}}
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  name: {{ template "trackableappname" . }}
-  annotations:
-    {{ if .Values.gitlab.app }}app.gitlab.com/app: {{ .Values.gitlab.app | quote }}{{ end }}
-    {{ if .Values.gitlab.env }}app.gitlab.com/env: {{ .Values.gitlab.env | quote }}{{ end }}
-  labels:
-    app: {{ template "appname" . }}
-    track: "{{ .Values.application.track }}"
-    tier: "{{ .Values.application.tier }}"
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version | replace "+" "_" }}"
-    release: {{ .Release.Name }}
-    heritage: {{ .Release.Service }}
-spec:
-  replicas: {{ .Values.replicaCount }}
-  template:
-    metadata:
-      annotations:
-        checksum/application-secrets: "{{ .Values.application.secretChecksum }}"
-        {{ if .Values.gitlab.app }}app.gitlab.com/app: {{ .Values.gitlab.app | quote }}{{ end }}
-        {{ if .Values.gitlab.env }}app.gitlab.com/env: {{ .Values.gitlab.env | quote }}{{ end }}
-{{- if .Values.podAnnotations }}
-{{ toYaml .Values.podAnnotations | indent 8 }}
-{{- end }}
-      labels:
-        app: {{ template "appname" . }}
-        track: "{{ .Values.application.track }}"
-        tier: "{{ .Values.application.tier }}"
-        release: {{ .Release.Name }}
-    spec:
-      imagePullSecrets:
-{{ toYaml .Values.image.secrets | indent 10 }}
-      containers:
-      - name: {{ .Chart.Name }}
-        image: "{{ .Values.image.repository }}:{{ .Values.image.tag }}"
-        imagePullPolicy: {{ .Values.image.pullPolicy }}
-        {{- if .Values.application.secretName }}
-        envFrom:
-        - secretRef:
-            name: {{ .Values.application.secretName }}
-        {{- end }}
-        env:
-        - name: DATABASE_URL
-          value: {{ .Values.application.database_url | quote }}
-        ports:
-        - name: "{{ .Values.service.name }}"
-          containerPort: {{ .Values.service.internalPort }}
-        livenessProbe:
-          httpGet:
-            path: {{ .Values.livenessProbe.path }}
-            port: {{ .Values.service.internalPort }}
-            scheme: {{ .Values.livenessProbe.scheme }}
-          initialDelaySeconds: {{ .Values.livenessProbe.initialDelaySeconds }}
-          timeoutSeconds: {{ .Values.livenessProbe.timeoutSeconds }}
-        readinessProbe:
-          httpGet:
-            path: {{ .Values.readinessProbe.path }}
-            port: {{ .Values.service.internalPort }}
-            scheme: {{ .Values.readinessProbe.scheme }}
-          initialDelaySeconds: {{ .Values.readinessProbe.initialDelaySeconds }}
-          timeoutSeconds: {{ .Values.readinessProbe.timeoutSeconds }}
-        resources:
-{{ toYaml .Values.resources | indent 12 }}
-{{- end -}}
