--- auto-deploy-app/templates/hpa.yaml
@@ -1,19 +1,0 @@
-{{- if and .Values.hpa.enabled .Values.resources.requests -}}
-apiVersion: autoscaling/v1
-kind: HorizontalPodAutoscaler
-metadata:
-  name: {{ template "fullname" . }}
-  labels:
-    app: {{ template "appname" . }}
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version | replace "+" "_" }}"
-    release: {{ .Release.Name }}
-    heritage: {{ .Release.Service }}
-spec:
-  scaleTargetRef:
-    kind: Deployment
-    name: {{ template "appname" . }}
-    apiVersion: apps/v1beta1
-  minReplicas: {{ .Values.hpa.minReplicas }}
-  maxReplicas: {{ .Values.hpa.maxReplicas }}
-  targetCPUUtilizationPercentage: {{ .Values.hpa.targetCPUUtilizationPercentage }}
-{{- end -}}
