--- auto-deploy-app/values.yaml
@@ -1,70 +1,0 @@
-# Default values for chart.
-# This is a YAML-formatted file.
-# Declare variables to be passed into your templates.
-replicaCount: 1
-image:
-  repository: gitlab.example.com/group/project
-  tag: stable
-  pullPolicy: Always
-  secrets:
-    - name: gitlab-registry
-podAnnotations: {}
-application:
-  track: stable
-  tier: web
-  migrateCommand:
-  initializeCommand:
-  secretName:
-  secretChecksum:
-hpa:
-  enabled: false
-  minReplicas: 1
-  maxReplicas: 5
-  targetCPUUtilizationPercentage: 80
-gitlab:
-  app:
-  env:
-service:
-  enabled: true
-  annotations: {}
-  name: web
-  type: ClusterIP
-  url: http://my.host.com/
-  additionalHosts:
-  commonName:
-  externalPort: 5000
-  internalPort: 5000
-ingress:
-  tls:
-    enabled: true
-    secretName: ""
-  annotations:
-    kubernetes.io/tls-acme: "true"
-    kubernetes.io/ingress.class: "nginx"
-livenessProbe:
-  path: "/"
-  initialDelaySeconds: 15
-  timeoutSeconds: 15
-  scheme: "HTTP"
-readinessProbe:
-  path: "/"
-  initialDelaySeconds: 5
-  timeoutSeconds: 3
-  scheme: "HTTP"
-postgresql:
-  enabled: true
-resources:
-#  limits:
-#    cpu: 100m
-#    memory: 128Mi
-  requests:
-#    cpu: 100m
-#    memory: 128Mi
-
-## Configure PodDisruptionBudget
-## ref: https://kubernetes.io/docs/concepts/workloads/pods/disruptions/
-#
-podDisruptionBudget:
-  enabled: false
-  # minAvailable: 1
-  maxUnavailable: 1
