+++ internal/pkg/templating/controller.tmpl
