+++ internal/pkg/templating/controllerFunc.tmpl
