+++ internal/pkg/templating/resourceFunc.tmpl
