+++ internal/pkg/templating/resources.go
